# packages/api-contracts/openapi/backend.openapi.json
{
  "components": {
    "schemas": {
...
      "CreateControlAccount": {
        "additionalProperties": false,
        "properties": {
          "displayName": {
            "type": "string"
          },
          "initialPassword": {
            "format": "password",
            "maxLength": 256,
            "minLength": 12,
            "type": "string",
            "writeOnly": true
          },
          "loginName": {
            "type": "string"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "maxItems": 64,
            "minItems": 0,
            "type": "array"
          }
        },
        "required": [
          "displayName",
          "initialPassword",
          "loginName",
          "permissions"
        ],
        "type": "object"
      },
...
      "Failure": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "success": {
            "const": false,
            "type": "boolean"
          },
          "traceId": {
            "pattern": "^[0-9a-f]{32}$",
            "type": "string"
          }
        },
        "required": [
          "error",
          "success",
          "traceId"
        ],
        "type": "object"
      },
...
      "SuccessControlMutation": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ControlMutation"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "const": true,
            "type": "boolean"
          },
          "traceId": {
            "pattern": "^[0-9a-f]{32}$",
            "type": "string"
          }
        },
        "required": [
          "data",
          "success",
          "traceId"
        ],
        "type": "object"
      },
...
      "SuccessPageResponseControlAccountView": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageResponseControlAccountView"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "const": true,
            "type": "boolean"
          },
          "traceId": {
            "pattern": "^[0-9a-f]{32}$",
            "type": "string"
          }
        },
        "required": [
          "data",
          "success",
          "traceId"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CustomerToken": {
        "description": "Bearer加一个空格再加原始opaque Token；仅客户小程序，无客户Cookie",
        "in": "header",
        "name": "X-Customer-Token",
        "type": "apiKey"
      },
      "PlatformCookie": {
        "description": "仅PLATFORM WEB；写请求必须独立CSRF及同源来源",
        "in": "cookie",
        "name": "__Secure-pet_platform_sid",
        "type": "apiKey"
      },
      "StaffCookie": {
        "description": "生产Cookie；local HTTP使用pet_dev_staff_sid，写请求另需X-CSRF-Token及固定Origin/Referer",
        "in": "cookie",
        "name": "__Secure-pet_staff_sid",
        "type": "apiKey"
      },
      "StaffToken": {
        "description": "Bearer加一个空格再加原始opaque Token；与Cookie不能并存",
        "in": "header",
        "name": "X-Staff-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "企业应用公共接口契约",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/platform/accounts": {
      "get": {
        "description": "platform:account:list；筛选分页排序，id稳定次序，total字符串。",
        "operationId": "listControlAccounts",
        "parameters": [
          {
            "in": "query",
            "name": "keyword"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "enum": [
                "ACTIVE",
                "DISABLED"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "maximum": 2147483647,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sortBy"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessPageResponseControlAccountView"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "参数格式或未知字段"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "未登录或平台身份失效"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "权限、授予上限或CSRF拒绝"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "版本、状态或最后管理入口冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "字段校验失败"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "敏感操作频控"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "依赖故障，拒绝执行"
          }
        },
        "security": [
          {
            "PlatformCookie": []
          }
        ],
        "summary": "平台账号列表",
        "tags": [
          "control-management-controller"
        ]
      },
      "post": {
        "description": "platform:account:create + platform:account:grant；正式PLATFORM身份，写请求CSRF，版本和治理约束在事务中复核。",
        "operationId": "createControlAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateControlAccount"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessControlMutation"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "参数格式或未知字段"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "未登录或平台身份失效"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "权限、授予上限或CSRF拒绝"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "版本、状态或最后管理入口冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "字段校验失败"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "敏感操作频控"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "依赖故障，拒绝执行"
          }
        },
        "security": [
          {
            "PlatformCookie": []
          }
        ],
        "summary": "创建平台账号与权限",
        "tags": [
          "control-management-controller"
        ]
      }
    },
    "/api/platform/accounts/{id}": {
      "get": {
        "description": "platform:account:detail；正式PLATFORM身份，写请求CSRF，版本和治理约束在事务中复核。",
        "operationId": "getControlAccount",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessControlAccountView"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "参数格式或未知字段"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "未登录或平台身份失效"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "权限、授予上限或CSRF拒绝"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "版本、状态或最后管理入口冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "字段校验失败"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "敏感操作频控"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "依赖故障，拒绝执行"
          }
        },
        "security": [
          {
            "PlatformCookie": []
          }
        ],
        "summary": "平台账号详情",
        "tags": [
          "control-management-controller"
        ]
      },
      "put": {
        "description": "platform:account:update；正式PLATFORM身份，写请求CSRF，版本和治理约束在事务中复核。",
        "operationId": "editControlAccount",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditControlAccount"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessControlMutation"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "参数格式或未知字段"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "未登录或平台身份失效"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "权限、授予上限或CSRF拒绝"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "版本、状态或最后管理入口冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "字段校验失败"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "敏感操作频控"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "依赖故障，拒绝执行"
          }
        },
        "security": [
          {
            "PlatformCookie": []
          }
        ],
        "summary": "修改平台账号基本资料",
        "tags": [
          "control-management-controller"
        ]
      }
    },
    "/api/platform/accounts/{id}/password": {
      "put": {
        "description": "platform:account:reset-password；正式PLATFORM身份，写请求CSRF，版本和治理约束在事务中复核。",
        "operationId": "resetControlAccountPassword",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlAccountReset"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessControlMutation"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "参数格式或未知字段"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "未登录或平台身份失效"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "权限、授予上限或CSRF拒绝"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "版本、状态或最后管理入口冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "字段校验失败"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "敏感操作频控"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "依赖故障，拒绝执行"
          }
        },
        "security": [
          {
            "PlatformCookie": []
          }
        ],
        "summary": "重置平台账号密码",
        "tags": [
          "control-management-controller"
        ]
      }
    },
    "/api/platform/accounts/{id}/permissions": {
      "put": {
        "description": "platform:account:grant；正式PLATFORM身份，写请求CSRF，版本和治理约束在事务中复核。",
        "operationId": "setControlAccountGrants",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlAccountGrants"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessControlMutation"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "参数格式或未知字段"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "未登录或平台身份失效"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "权限、授予上限或CSRF拒绝"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "版本、状态或最后管理入口冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "字段校验失败"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "敏感操作频控"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "依赖故障，拒绝执行"
          }
        },
        "security": [
          {
            "PlatformCookie": []
          }
        ],
        "summary": "配置平台账号权限",
        "tags": [
          "control-management-controller"
        ]
      }
    },
...
  }
}